server/requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.14.0
bcrypt==4.2.1
PyJWT==2.10.1
orjson==3.10.12
APScheduler==3.10.4
